Adds an operator manifest example for running on Kubernetes (#17)
* Adds an operator manifest example for running on Kubernetes

This new operator.yaml will become the template for a versioned release asset
when we start cutting releases.  That way we can include the fully-specified
`PrefectHQ/prefect-operator` version tag in a place that people can permanently
reference.

This change also changes the SQLite version of the server from a `StatefulSet`
to a `Deployment` with an explicit PVC, fixed replicas, and the `Recreate`
rollout strategy.  This makes it easier to reason about what the server
deployment will look like and simplifies the code.

Closes #8

* Typo