Adds status and conditions for `PrefectServer` and `PrefectWorkPool` (#34)
* Adds status and conditions for `PrefectServer` and `PrefectWorkPool`

It's standard practice with Kubernetes controllers to maintain some status
fields describing the observed state of the resource, as well as conditions to
describe some orthogonal sub-statuses.  Here, we add the observed version of
Prefect for both, and how many workers are running.  For conditions, we maintain
one for each of the components (PVC, migration Job, Deployment, and Service)
for the Prefect server, and just the Deployment of the work pool as a starting
point.

Closes #33

[1]: https://maelvls.dev/kubernetes-conditions/
[2]: https://sdk.operatorframework.io/docs/building-operators/golang/advanced-topics/
[3]: https://book-v1.book.kubebuilder.io/basics/status_subresource