Adds `Ready` status boolean for `PrefectServer` and `PrefectWorkPool` (#50)
* Restores CRD descriptions

In our previous change to de-scaffold the extra operator-sdk boilerplate, we
lost the descriptions in our CRDs.  Since we're not experiencing the issue
described in the comment, we can keep our descriptions for now to help with
documentation.

* Adds `Ready` status boolean for `PrefectServer` and `PrefectWorkPool`

For both of these, `Ready` will be true when there is at least one ready
replica of the underlying deployment.  While working on this, I also noticed
that we were _always_ attempting to add a condition, even if nothing had
changed (and thus the condition struct was empty), which would lead to quiet
errors with all further reconciliations.  Now we'll exit early if the condition
isn't populated, report errors if we can't update the status, and explicitly
update the status when non-conditions are changed.

Fixes #49

* Adding unit tests for readiness