Support specifying which namespace(s) to watch (#61)
* Support specifying which namespace(s) to watch

By default, an Operator is configured to watch objects from any
namespace. Its access is only limited by RBAC.

The Prefect Operator should only need to manage resources in the same
namespace where it is deployed. So let's default to restricting it to
its own namespace.

References:
- https://book.kubebuilder.io/cronjob-tutorial/empty-main.html
- https://sdk.operatorframework.io/docs/building-operators/golang/operator-scope

* Fix environment variable name

Co-authored-by: Chris Guidry <chris@theguidrys.us>

* Disable namespace-scoped setting for now

Disables the namespace-scoped setting (the WATCH_NAMESPACES env var)
to keep the Operator cluser-scoped for now. The setting will be there in
the future if we ever want to scope it down.

When/if we do, we should look into scoping the RBAC down so it isn't
cluster-scoped.

---------

Co-authored-by: Chris Guidry <chris@theguidrys.us>