CI: always push short sha image tags
Follows the pattern we use in the prefect-cloud-operator,
always pushing the short sha. This is currently used for updating the
version running in dev but will likely be handy in the future to refer
to specific image versions as well.

Also:
- adds comments explaining what each tag spec does
- adds the 'latest' tag for main merges

Related to https://linear.app/prefect/issue/PLA-343/automatically-update-prefect-operator-helm-chart-versions-in-ccd-to