Fixes Prefect image command-line arguments and labels (#140)
We were passing `Command` to the Prefect `server` container, which in
K8s would bypass the configured `CMD` of the Prefect Docker image.  We
want to preserve the Prefect entrypoint and pass our command as the
_arguments_ to that entrypoint.  This allows us to do things like pass
`EXTRA_PIP_PACKAGES` and have the Prefect entrypoint script install
them.

We were also applying the server's version as a label on the deployment
and its pods, but those labels are immutable after creation.  This means
we can't update the version of the Prefect server.  We don't need the
version stamped as a label, as we will have the version definitively in
the `Image` of the container.  If we find we need it as metadata, we can
put it in an annotation later.